Convert inner classes of MapReplicationOperation to IdentifiedDataSerializable
